sink: attempt to make image tag unique per push
According to the jenkins doc [1] I found the environment variable `sha1`
should be set. This patch uses a shortened version to add to the tag
when possible so that test runs on the same PR but different commits
(pushes) done before the first test run are less likely to interfere
with each other.

[1] https://plugins.jenkins.io/ghprb/

Signed-off-by: John Mulligan <jmulligan@redhat.com>