jobs/scripts: Pre-fetch vagrant boxes for CentOS Stream
We started seeing the signature error mentioned in the following bug
report while executing `dnf update`:

https://bugzilla.redhat.com/show_bug.cgi?id=1918777

Due to the presence of an old pre-installed version of RPM in vagrant
box for CentOS Stream 8 we were unable to perform a complete system
update. Therefore make sure that we keep up with the recent versions
of libvirt vagrant boxes for CentOS Stream.